+++ src/cli.ts
@@ -2,11 +2,13 @@
 import { Command } from "commander";
 import { TRACEPACK_VERSION } from "./core/manifest.js";
 import { registerAssert } from "./commands/assert.js";
+import { registerClean } from "./commands/clean.js";
 import { registerDoctor } from "./commands/doctor.js";
 import { registerFinish } from "./commands/finish.js";
 import { registerReport } from "./commands/report.js";
 import { registerRun } from "./commands/run.js";
 import { registerStart } from "./commands/start.js";
+import { registerStatus } from "./commands/status.js";
 
 const program = new Command();
 
@@ -17,9 +19,11 @@ program
 
 registerStart(program);
 registerRun(program);
+registerStatus(program);
 registerFinish(program);
 registerReport(program);
 registerAssert(program);
+registerClean(program);
 registerDoctor(program);
 
 program.showHelpAfterError();